ci: use remote speculos for LWD tests
