+++ e2e/desktop/tests/specs/provider.swap.spec.ts
@@ -1,0 +1,263 @@
+import test from "tests/fixtures/common";
+import { Account, TokenAccount } from "@ledgerhq/live-common/e2e/enum/Account";
+import { AppInfos } from "@ledgerhq/live-common/e2e/enum/AppInfos";
+import { setExchangeDependencies } from "@ledgerhq/live-common/e2e/speculos";
+import { Swap } from "@ledgerhq/live-common/e2e/models/Swap";
+import { addBugLink, addTmsLink } from "tests/utils/allureUtils";
+import { getDescription } from "tests/utils/customJsonReporter";
+import { Provider } from "@ledgerhq/live-common/e2e/enum/Provider";
+import { setupEnv, performSwapUntilQuoteSelectionStep } from "tests/utils/swapUtils";
+import { liveDataWithAddressCommand } from "tests/utils/cliCommandsUtils";
+
+const app: AppInfos = AppInfos.EXCHANGE;
+
+test.describe("Swap - Provider redirection", () => {
+  const fromAccount = Account.ETH_1;
+  const toAccount = TokenAccount.ETH_USDC_1;
+  const provider = Provider.VELORA;
+  setupEnv(true);
+
+  const accPair: string[] = [fromAccount, toAccount].map(acc =>
+    acc.currency.speculosApp.name.replace(/ /g, "_"),
+  );
+
+  test.beforeEach(async () => {
+    setExchangeDependencies(
+      accPair.map(appName => ({
+        name: appName,
+      })),
+    );
+  });
+
+  test.use({
+    userdata: "skip-onboarding",
+    speculosApp: app,
+
+    cliCommandsOnApp: [
+      [
+        {
+          app: fromAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(fromAccount),
+        },
+        {
+          app: toAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(toAccount),
+        },
+      ],
+      { scope: "test" },
+    ],
+  });
+
+  test(
+    `Swap test provider redirection (${provider.uiName})`,
+    {
+      tag: ["@NanoSP", "@LNS", "@NanoX", "@Stax", "@Flex", "@NanoGen5", "@ethereum", "@family-evm"],
+      annotation: {
+        type: "TMS",
+        description: "B2CQA-3119",
+      },
+    },
+    async ({ app, electronApp }) => {
+      await addTmsLink(getDescription(test.info().annotations, "TMS").split(", "));
+
+      const minAmount = await app.swap.getMinimumAmount(fromAccount, toAccount);
+      const swap = new Swap(fromAccount, toAccount, minAmount);
+
+      await performSwapUntilQuoteSelectionStep(app, electronApp, swap, minAmount);
+
+      await app.swap.selectSpecificProvider(provider, electronApp);
+      await app.swap.goToProviderLiveApp(electronApp, provider.uiName);
+      await app.swap.verifyProviderURL(electronApp, provider.uiName, swap);
+      await app.liveApp.verifyLiveAppTitle(provider.uiName.toLowerCase());
+    },
+  );
+});
+
+test.describe("Swap - 1inch flow", () => {
+  const fromAccount = Account.ETH_1;
+  const toAccount = TokenAccount.ETH_USDT_1;
+  const provider = Provider.ONE_INCH;
+  setupEnv(true);
+
+  const accPair: string[] = [fromAccount, toAccount].map(acc =>
+    acc.currency.speculosApp.name.replace(/ /g, "_"),
+  );
+
+  test.beforeEach(async () => {
+    setExchangeDependencies(
+      accPair.map(appName => ({
+        name: appName,
+      })),
+    );
+  });
+
+  test.use({
+    userdata: "skip-onboarding",
+    speculosApp: app,
+
+    cliCommandsOnApp: [
+      [
+        {
+          app: fromAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(fromAccount),
+        },
+        {
+          app: toAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(toAccount),
+        },
+      ],
+      { scope: "test" },
+    ],
+  });
+
+  test(
+    `Swap test 1inch flow (${provider.uiName})`,
+    {
+      tag: ["@NanoSP", "@LNS", "@NanoX", "@Stax", "@Flex", "@NanoGen5", "@ethereum", "@family-evm"],
+      annotation: [
+        {
+          type: "TMS",
+          description: "B2CQA-3120",
+        },
+        { type: "BUG", description: "QAA-854" },
+      ],
+    },
+    async ({ app, electronApp }) => {
+      await addTmsLink(getDescription(test.info().annotations, "TMS").split(", "));
+      await addBugLink(getDescription(test.info().annotations, "BUG").split(", "));
+
+      const minAmount = await app.swap.getMinimumAmount(fromAccount, toAccount);
+      const swap = new Swap(fromAccount, toAccount, minAmount);
+
+      await performSwapUntilQuoteSelectionStep(app, electronApp, swap, minAmount);
+
+      await app.swap.selectSpecificProvider(provider, electronApp);
+      await app.swap.clickExchangeButton(electronApp);
+      await app.swap.checkElementsPresenceOnSwapApprovalStep(electronApp);
+      await app.swap.clickExecuteSwapButton(electronApp);
+      await app.swap.clickContinueButton();
+    },
+  );
+});
+
+test.describe("Swap - Check Best Offer", () => {
+  const fromAccount = Account.ETH_1;
+  const toAccount = Account.BTC_NATIVE_SEGWIT_1;
+  setupEnv(true);
+
+  test.beforeEach(async () => {
+    const accountPair: string[] = [fromAccount, toAccount].map(acc =>
+      acc.currency.speculosApp.name.replace(/ /g, "_"),
+    );
+    setExchangeDependencies(accountPair.map(name => ({ name })));
+  });
+
+  test.use({
+    userdata: "skip-onboarding",
+    speculosApp: app,
+
+    cliCommandsOnApp: [
+      [
+        {
+          app: fromAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(fromAccount),
+        },
+        {
+          app: toAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(toAccount),
+        },
+      ],
+      { scope: "test" },
+    ],
+  });
+
+  test(
+    `Swap ${fromAccount.currency.name} to ${toAccount.currency.name} - Check "Best Offer"`,
+    {
+      tag: [
+        "@NanoSP",
+        "@LNS",
+        "@NanoX",
+        "@Stax",
+        "@Flex",
+        "@NanoGen5",
+        "@ethereum",
+        "@family-evm",
+        "@bitcoin",
+        "@family-bitcoin",
+      ],
+      annotation: { type: "TMS", description: "B2CQA-2327" },
+    },
+    async ({ app, electronApp }) => {
+      await addTmsLink(getDescription(test.info().annotations, "TMS").split(", "));
+
+      const minAmount = await app.swap.getMinimumAmount(fromAccount, toAccount);
+      const swap = new Swap(fromAccount, toAccount, minAmount);
+
+      await performSwapUntilQuoteSelectionStep(app, electronApp, swap, minAmount);
+      await app.swap.selectExchangeWithoutKyc(electronApp);
+      await app.swap.checkBestOffer(electronApp);
+    },
+  );
+});
+
+test.describe("Swap - Landing page", () => {
+  const fromAccount = Account.ETH_1;
+  const toAccount = TokenAccount.ETH_USDC_1;
+
+  setupEnv(true);
+
+  test.beforeEach(async () => {
+    const accountPair: string[] = [fromAccount, toAccount].map(acc =>
+      acc.currency.speculosApp.name.replace(/ /g, "_"),
+    );
+    setExchangeDependencies(accountPair.map(name => ({ name })));
+  });
+
+  test.use({
+    userdata: "skip-onboarding",
+    speculosApp: app,
+
+    cliCommandsOnApp: [
+      [
+        {
+          app: fromAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(fromAccount),
+        },
+        {
+          app: toAccount.currency.speculosApp,
+          cmd: liveDataWithAddressCommand(toAccount),
+        },
+      ],
+      { scope: "test" },
+    ],
+  });
+
+  test(
+    `Swap landing page`,
+    {
+      tag: ["@NanoSP", "@LNS", "@NanoX", "@Stax", "@Flex", "@NanoGen5", "@ethereum", "@family-evm"],
+      annotation: { type: "TMS", description: "B2CQA-2918" },
+    },
+    async ({ app, electronApp }) => {
+      await addTmsLink(getDescription(test.info().annotations, "TMS").split(", "));
+
+      const minAmount = await app.swap.getMinimumAmount(fromAccount, toAccount);
+
+      if (!minAmount) {
+        throw new Error("Test failed: No quotes retrieved from swap API.");
+      }
+
+      const swap = new Swap(fromAccount, toAccount, minAmount);
+
+      await performSwapUntilQuoteSelectionStep(app, electronApp, swap, minAmount);
+      const providerList = await app.swap.getProviderList(electronApp);
+      await app.swap.checkQuotesContainerInfos(
+        electronApp,
+        providerList,
+        toAccount.currency.ticker,
+      );
+      await app.swap.checkBestOffer(electronApp);
+    },
+  );
+});
