+++ e2e/desktop/tests/specs/send.swap.spec.ts
