+++ e2e/desktop/tests/utils/allureUtils.ts
@@ -9,6 +9,20 @@ import { isLastRetry } from "tests/utils/testInfoUtils";
 const readFileAsync = promisify(readFile);
 const IS_NOT_MOCK = process.env.MOCK == "0";
 
+async function attachIfExists(
+  testInfo: TestInfo,
+  name: string,
+  filePath: string,
+  contentType: string,
+) {
+  try {
+    const data = await readFileAsync(filePath);
+    await testInfo.attach(name, { body: data, contentType });
+  } catch {
+    // File does not exist → silently ignore
+  }
+}
+
 export async function addTmsLink(ids: string[]) {
   for (const id of ids) {
     await allure.tms(id);
@@ -48,6 +62,13 @@ export async function captureArtifacts(
       path: filePath,
       contentType: "application/json",
     });
+
+    await attachIfExists(
+      testInfo,
+      "Network failures",
+      testInfo.outputPath("network.log"),
+      "text/plain",
+    );
   }
 
   const video = page.video();
